Gift Ai assitante/requirements.txt
flask
flask-cors
google-generativeai
groq
python-dotenv
requests
